menu api completed

+++ server/src/main/java/com/example/restaurant/repositry/UserRepositry.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRepositry extends JpaRepository<User, Long> {
+public interface UserRepositry extends JpaRepository<User, Integer> {
 }
